RUMM-271 Rename artifacts
