+++ build.gradle.kts
@@ -57,14 +57,14 @@ gitDiffTask("unitTestChanged") {
         ":sample:kotlin-timber:assembleDebug"
     )
     dependsOnDiff(
-        "dd-sdk-androidx-fragments/.*",
-        ":dd-sdk-androidx-fragments:testDebugUnitTest",
-        ":dd-sdk-androidx-fragments:testReleaseUnitTest"
+        "dd-sdk-androidx-fragment/.*",
+        ":dd-sdk-androidx-fragment:testDebugUnitTest",
+        ":dd-sdk-androidx-fragment:testReleaseUnitTest"
     )
     dependsOnDiff(
-        "dd-sdk-support-fragments/.*",
-        ":dd-sdk-support-fragments:testDebugUnitTest",
-        ":dd-sdk-support-fragments:testReleaseUnitTest"
+        "dd-sdk-support-fragment/.*",
+        ":dd-sdk-support-fragment:testDebugUnitTest",
+        ":dd-sdk-support-fragment:testReleaseUnitTest"
     )
     dependsOnDiff(
         "tools/detekt/.*",
@@ -126,8 +126,8 @@ tasks.register("ktlintCheckAll") {
     dependsOn(
         ":dd-sdk-android:ktlintCheck",
         ":dd-sdk-android-timber:ktlintCheck",
-        ":dd-sdk-androidx-fragments:ktlintCheck",
-        ":dd-sdk-support-fragments:ktlintCheck",
+        ":dd-sdk-androidx-fragment:ktlintCheck",
+        ":dd-sdk-support-fragment:ktlintCheck",
         ":instrumented:integration:ktlintCheck",
         ":instrumented:benchmark:ktlintCheck",
         ":tools:detekt:ktlintCheck",
@@ -141,10 +141,10 @@ tasks.register("lintCheckAll") {
         ":dd-sdk-android:lintRelease",
         ":dd-sdk-android-timber:lintDebug",
         ":dd-sdk-android-timber:lintRelease",
-        ":dd-sdk-androidx-fragments:lintDebug",
-        ":dd-sdk-androidx-fragments:lintRelease",
-        ":dd-sdk-support-fragments:lintDebug",
-        ":dd-sdk-support-fragments:lintRelease"
+        ":dd-sdk-androidx-fragment:lintDebug",
+        ":dd-sdk-androidx-fragment:lintRelease",
+        ":dd-sdk-support-fragment:lintDebug",
+        ":dd-sdk-support-fragment:lintRelease"
     )
 }
 
@@ -152,8 +152,8 @@ tasks.register("detektAll") {
     dependsOn(
         ":dd-sdk-android:detekt",
         ":dd-sdk-android-timber:detekt",
-        ":dd-sdk-androidx-fragments:detekt",
-        ":dd-sdk-support-fragments:detekt",
+        ":dd-sdk-androidx-fragment:detekt",
+        ":dd-sdk-support-fragment:detekt",
         ":instrumented:integration:detekt",
         ":instrumented:benchmark:detekt",
         ":tools:unit:detekt"
@@ -166,10 +166,10 @@ tasks.register("jacocoReportAll") {
         ":dd-sdk-android:jacocoTestReleaseUnitTestReport",
         ":dd-sdk-android-timber:jacocoTestDebugUnitTestReport",
         ":dd-sdk-android-timber:jacocoTestReleaseUnitTestReport",
-        ":dd-sdk-androidx-fragments:jacocoTestDebugUnitTestReport",
-        ":dd-sdk-androidx-fragments:jacocoTestReleaseUnitTestReport",
-        ":dd-sdk-support-fragments:jacocoTestDebugUnitTestReport",
-        ":dd-sdk-support-fragments:jacocoTestReleaseUnitTestReport",
+        ":dd-sdk-androidx-fragment:jacocoTestDebugUnitTestReport",
+        ":dd-sdk-androidx-fragment:jacocoTestReleaseUnitTestReport",
+        ":dd-sdk-support-fragment:jacocoTestDebugUnitTestReport",
+        ":dd-sdk-support-fragment:jacocoTestReleaseUnitTestReport",
         ":tools:detekt:jacocoTestReport",
         ":tools:unit:jacocoTestDebugUnitTestReport",
         ":tools:unit:jacocoTestReleaseUnitTestReport"
