+++ dd-sdk-androidx-fragment/build.gradle.kts
