+++ dd-sdk-androidx-fragment/src/main/kotlin/com/datadog/android/androidx/fragment/CompatFragmentLifecycleCallbacks.kt
@@ -1,0 +1,17 @@
+package com.datadog.android.androidx.fragment
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import com.datadog.android.rum.GlobalRum
+
+internal object CompatFragmentLifecycleCallbacks : FragmentManager.FragmentLifecycleCallbacks() {
+    override fun onFragmentResumed(fm: FragmentManager, f: Fragment) {
+        super.onFragmentResumed(fm, f)
+        GlobalRum.get().startView(f, f.javaClass.canonicalName!!)
+    }
+
+    override fun onFragmentPaused(fm: FragmentManager, f: Fragment) {
+        super.onFragmentPaused(fm, f)
+        GlobalRum.get().stopView(f)
+    }
+}
