+++ dd-sdk-androidx-fragment/src/test/kotlin/com/datadog/android/androidx/fragment/DefaultFragmentLifecycleCallbacksTest.kt
@@ -1,0 +1,57 @@
+package com.datadog.android.androidx.fragment
+
+import android.app.Fragment
+import com.nhaarman.mockitokotlin2.eq
+import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.verify
+import fr.xgouchet.elmyr.Forge
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
+import org.junit.jupiter.api.extension.Extensions
+import org.mockito.Mock
+import org.mockito.junit.jupiter.MockitoExtension
+import org.mockito.junit.jupiter.MockitoSettings
+import org.mockito.quality.Strictness
+
+@Extensions(
+    ExtendWith(MockitoExtension::class)
+)
+@MockitoSettings(strictness = Strictness.LENIENT)
+internal class DefaultFragmentLifecycleCallbacksTest : LifecycleCallbacksTest() {
+    lateinit var underTest: DefaultFragmentLifecycleCallbacks
+
+    @Mock
+    lateinit var mockFragment: TestFragment1
+
+    @BeforeEach
+    override fun `set up`(forge: Forge) {
+        super.`set up`(forge)
+        underTest = DefaultFragmentLifecycleCallbacks
+    }
+
+    @Test
+    fun `when fragment resumed it will start a view event`(forge: Forge) {
+        // when
+        underTest.onFragmentResumed(mock(), mockFragment)
+        // then
+        verify(mockRumMonitor).startView(
+            eq(mockFragment),
+            eq(mockFragment::class.java.canonicalName!!),
+            eq(emptyMap())
+        )
+    }
+
+    @Test
+    fun `when fragment paused it will stop a view event`(forge: Forge) {
+        // when
+        underTest.onFragmentPaused(mock(), mockFragment)
+        // then
+        verify(mockRumMonitor).stopView(
+            eq(mockFragment),
+            eq(emptyMap())
+        )
+    }
+
+    internal open class TestFragment1 : Fragment()
+}
