--- dd-sdk-androidx-fragments/src/main/AndroidManifest.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
-  ~ This product includes software developed at Datadog (https://www.datadoghq.com/).
-  ~ Copyright 2016-2019 Datadog, Inc.
-  -->
-
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.datadog.android.androidx.fragments">
-
-    <application/>
-
-</manifest>
