+++ dd-sdk-support-fragment/build.gradle.kts
