+++ dd-sdk-support-fragment/src/main/kotlin/com/datadog/android/support/fragment/CompatFragmentLifecycleCallbacks.kt
@@ -1,0 +1,17 @@
+package com.datadog.android.support.fragment
+
+import android.support.v4.app.Fragment
+import android.support.v4.app.FragmentManager
+import com.datadog.android.rum.GlobalRum
+
+internal object CompatFragmentLifecycleCallbacks : FragmentManager.FragmentLifecycleCallbacks() {
+    override fun onFragmentResumed(fm: FragmentManager, f: Fragment) {
+        super.onFragmentResumed(fm, f)
+        GlobalRum.get().startView(f, f.javaClass.canonicalName!!)
+    }
+
+    override fun onFragmentPaused(fm: FragmentManager, f: Fragment) {
+        super.onFragmentPaused(fm, f)
+        GlobalRum.get().stopView(f)
+    }
+}
