+++ sample/kotlin/src/main/kotlin/com/datadog/android/sample/SampleApplication.kt
@@ -9,7 +9,7 @@ import android.app.Application
 import android.util.Log
 import com.datadog.android.Datadog
 import com.datadog.android.DatadogConfig
-import com.datadog.android.androidx.fragments.TrackFragmentsAsViewsStrategy
+import com.datadog.android.androidx.fragment.TrackFragmentsAsViewsStrategy
 import com.datadog.android.rum.GlobalRum
 import com.datadog.android.rum.RumMonitor
 import com.datadog.android.rum.TrackActivitiesAsViewsStrategy
