+++ settings.gradle.kts
@@ -36,8 +36,8 @@ pluginManagement {
 
 include(":dd-sdk-android")
 include(":dd-sdk-android-timber")
-include(":dd-sdk-androidx-fragments")
-include(":dd-sdk-support-fragments")
+include(":dd-sdk-androidx-fragment")
+include(":dd-sdk-support-fragment")
 
 include(":instrumented:benchmark")
 include(":instrumented:integration")
